Fixed issue with packetReceived trigger causing crashes.
Due to some weird reasons with how minecraft works with threads and packets the old packetReceived trigger would sometimes cause crashes. This was solved by changing the injection location.
