+++ src/main/kotlin/com/chattriggers/ctjs/launch/plugin/CTJSTransformer.kt
@@ -34,7 +34,7 @@ class CTJSTransformer : BaseClassTransformer() {
             injectScreenshotHelper()
             injectNetHandlerPlayClient()
             injectRenderManager()
-            injectNetworkManager()
+            injectPacketThreadUtil()
             makeGuiScreenInjections()
 
             try {
